Fix the quantified apply functions.
The problem was that they were sharing a cache with the recursive
calls to the quantification function.  They need to share a node
cache, but they can't share memo tables otherwise you always end up
with a tautology (for ||, contradiction for &&).

Added an additional key to the memo tables to separate out recursive
invocations from the outer evaluation contexts.  Since the inner
context is always the same operation, this should be fine.