Fix makeDAG to use unique IDs instead of variable numbers as graph labels.
It was using the variable number as the node name in graphviz, but
that is terribly wrong.  Variables can appear more than once in a BDD
and this was collapsing all instances inappropriately, leading to
visualizations that looked nothing like BDDs.

Instead, use the unique ID assigned to each BDD node.